requirements.txt
streamlit==1.26.0
tensorflow-cpu==2.12.0
protobuf==4.23.3
pandas==1.5.3
numpy==1.23.5
requests==2.31.0
matplotlib==3.7.2
joblib==1.3.2
ta==0.11.0
